Update README.md
I am working at the university with researchers and teachers to facilitate the dissemination of their work openly through various channels so that their communities work transparently to increase their visibility and impact. I am working openly to influence the research culture, to incentivise collaborative working and to build a more replicable and robust science.